Fix GAE truncation bootstrap, clear all -Wconversion warnings, refresh README
GAE no longer leaks advantage across episode boundaries. A time-limit
truncation is now distinguished from a true terminal: it bootstraps from
the real next-state value (captured before the env resets) and resets the
advantage carry so the next episode cannot flow back. True terminals still
cut the bootstrap to zero. Applied on both the single-env and vectorized
collection paths, with a new unit test covering the truncation case.

Also cast all long long / size_t -> double divisions so the core, tools,
and tests build warning-clean under -Wall -Wextra -Wpedantic -Wshadow
-Wconversion (verified with GCC 13.3). README rewritten to document the
corrected GAE boundary handling, the warning-clean build, the recovery
config, simulator X11 prerequisites, and backend-specific config knobs.

+++ src/app/benchmark_main.cpp
@@ -60,7 +60,7 @@ double throughput(IntegratorType integ, long long steps) {
     }
     // Consume `sink` in a way the optimizer cannot prove is dead.
     if (sink == 1234567.89) DP_LOG_INFO("unreachable %.1f", sink);
-    return steps / w.seconds();
+    return static_cast<double>(steps) / w.seconds();
 }
 } // namespace
 
@@ -91,7 +91,7 @@ int main(int argc, char** argv) {
             StepResult sr = env.step(a);
             if (sr.terminal || sr.truncated) env.reset(123 + i);
         }
-        DP_LOG_INFO("Environment throughput: %.0f steps/sec", n / w.seconds());
+        DP_LOG_INFO("Environment throughput: %.0f steps/sec", static_cast<double>(n) / w.seconds());
     }
     return 0;
 }
